% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{build_event_type_choices}
\alias{build_event_type_choices}
\title{Build event type selectInput choices from the schema}
\usage{
build_event_type_choices(schema, iso)
}
\arguments{
\item{schema}{Loaded schema}

\item{iso}{ISO language code}
}
\value{
Named character vector for selectInput
}
\description{
Build event type selectInput choices from the schema
}
