% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{build_subtype_choices}
\alias{build_subtype_choices}
\title{Build subtype selectInput choices}
\usage{
build_subtype_choices(event_entry, iso)
}
\arguments{
\item{event_entry}{An event registry entry}

\item{iso}{ISO language code}
}
\value{
Named character vector for selectInput
}
\description{
Build subtype selectInput choices
}
