% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_json.R
\name{create_file_folder}
\alias{create_file_folder}
\title{Create a folder for a site-block combination}
\usage{
create_file_folder(site, block, base_folder = json_file_base_folder())
}
\arguments{
\item{site}{The site to create the folder for}

\item{block}{The block to create the folder for}

\item{base_folder}{Included for testing reasons, the default value should
otherwise be used}
}
\value{
TRUE if the directory was created successfully or already exists,
  FALSE otherwise.
}
\description{
Given a site and a block on that site, create a folder under
json_file_base_folder where the events.json file and related image files will
be stored. If the base folder doesn't exist, the function will throw an
error.
}
