% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{create_ui}
\alias{create_ui}
\title{Generate the UI for a list of elements in the structure file.}
\usage{
create_ui(widget_structure_list, ns)
}
\arguments{
\item{widget_structure_list}{The list of widget structures (from
ui_structure.json) to generate as UI}

\item{ns}{A namespacing function generated by shiny::NS to apply to the id's
of each generated widget}
}
\value{
A list of Shiny widgets
}
\description{
For a given list of widget structures as read from ui_structure.json, 
create_ui applies create_widget to each widget in the list.
Still used for app chrome rendering and by mod_table.R for legacy tables.
}
