% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file}
\usage{
delete_file(
  filepath,
  site = NULL,
  block = NULL,
  filepath_relative = FALSE,
  base_folder = json_file_base_folder()
)
}
\arguments{
\item{filepath}{The path to the file which should be deleted.}

\item{site}{The site where the event took place}

\item{block}{The block where the event took place}

\item{filepath_relative}{Set to TRUE and supply site and block if filepath is
relative to the events.json file.}

\item{base_folder}{Included for testing reasons}
}
\description{
Delete a file
}
