% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{determine_widget_type}
\alias{determine_widget_type}
\title{Determine the Shiny widget type from a schema property}
\usage{
determine_widget_type(prop)
}
\arguments{
\item{prop}{A resolved property descriptor}
}
\value{
A string: "selectInput", "numericInput", "dateInput", etc.
}
\description{
Determine the Shiny widget type from a schema property
}
