% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_form_fct_evaluate_js.R
\name{evaluate_condition}
\alias{evaluate_condition}
\title{Evalute some javascript conditions from ui_structure.json in R}
\usage{
evaluate_condition(js_condition, session)
}
\arguments{
\item{js_condition}{The javascript condition to evaluate as a string}

\item{session}{Current Shiny session in which to evaluate}
}
\value{
Returns either TRUE or FALSE. If the condition could not be
  evaluated, returns NULL.
}
\description{
Takes a condition written in javascript notation (visibility conditions in
ui_structure.json) and evaluates it in R.
}
\note{
Might not be best coding practice, but works as long as the
  js_condition doesn't have any typos. eval(parse(...)) is dangerous if it is
  used directly with user input, but here that is not the case. The user has
  no access to the ui_structure.json file.
}
