% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{extract_oneof_choices}
\alias{extract_oneof_choices}
\title{Extract selectInput choices from a oneOf array}
\usage{
extract_oneof_choices(one_of_array)
}
\arguments{
\item{one_of_array}{A list of oneOf entries each with const and title}
}
\value{
A list of choice descriptors with const, titles
}
\description{
Extract selectInput choices from a oneOf array
}
