% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{find_event_index}
\alias{find_event_index}
\title{Find the index of an event in a list}
\usage{
find_event_index(event, event_list)
}
\arguments{
\item{event}{The event whose index to identify}

\item{event_list}{The list of events where event is to be found}
}
\value{
The index if found, NULL otherwise
}
\description{
Find the first index corresponding to the given event in a list of events.
An event is considered equal to another if they have exactly the same
variables (though not necessarily in the same order) and these variables
have exactly the same values.
}
