% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_language.R
\name{get_category_names}
\alias{get_category_names}
\title{Find code and display names belonging to a given category}
\usage{
get_category_names(category, language = NULL)
}
\arguments{
\item{category}{The category (e.g. "variable_name") to find the names for}

\item{language}{(optional) The language of the display names}
}
\value{
A vector of code names. If language was supplied, the display names
  corresponding to the code names will be the names of the vector.
}
\description{
The categories are defined in the display_names.csv file. If language is
undefined, only code names will be returned. If a language is also supplied,
then the corresponding display names are set as the names of the vector of
code names.
}
