% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_event_list.R
\name{get_data_table}
\alias{get_data_table}
\title{Turn a list of events into a data frame}
\usage{
get_data_table(events, variable_names)
}
\arguments{
\item{events}{The list of events to turn into a data frame}

\item{variable_names}{The variables which should be displayed in the columns
of the data frame.}
}
\value{
A data frame with the events as rows and variable names as columns.
}
\description{
Takes a list of events and makes a data frame with given
  variables in columns. Also adds a column with the complete event list and a
  final column for ordering the list by date.
}
\note{
The function doesn't replace code names with display names. That is
  done separately so that when the app language is switched, we can change
  the table display names without having to create it again.
}
