% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_utils.R
\name{get_dynamic_rows_from_value}
\alias{get_dynamic_rows_from_value}
\title{Determine the dynamic rows based on row variable value}
\usage{
get_dynamic_rows_from_value(variable, value)
}
\value{
An atomic vector of rows, either option code names or numbers
}
\description{
This is used to go from the value of a variable determining the rows in a
dynamic row group to the rows themselves. If the row variable is a
selectInput, the rows equal the value, but if the row variable is a
numericInput, a vector of rows is generated instead
}
