% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{get_relevant_properties}
\alias{get_relevant_properties}
\title{Get properties relevant to the currently selected event/subtype}
\usage{
get_relevant_properties(schema, event_type, subtype = NULL)
}
\arguments{
\item{schema}{The loaded schema (from load_schema)}

\item{event_type}{The selected event type const}

\item{subtype}{The selected subtype const (or NULL)}
}
\value{
A list with: common, event_props, subtype_props, all
}
\description{
Get properties relevant to the currently selected event/subtype
}
