% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_utils.R
\name{get_table_variables}
\alias{get_table_variables}
\title{Find the variables whose value can be entered through a given table}
\usage{
get_table_variables(table_code_name)
}
\arguments{
\item{table_code_name}{The name of the table whose variables to fetch.}
}
\value{
A vector of variable names whose values are entered in a table.
}
\description{
Find the variables whose value can be entered through a given table
}
\note{
If a table has a dynamic row group whose rows are determined by an
  input widget's value, that widget's variable name will not be returned even
  though it could be read from the list returned by the table module.
}
