% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_utils.R
\name{get_variable_table}
\alias{get_variable_table}
\title{Find the table matching a variable name}
\usage{
get_variable_table(variable_name)
}
\arguments{
\item{variable_name}{The name of the variable of interest}
}
\value{
The code name of the table where the variable is entered, or NULL
  if not found.
}
\description{
If a variable's value is entered in a table, return the name of that table
}
