% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_event_format.R
\name{is_missing_value}
\alias{is_missing_value}
\title{Check which elements of an atomic vector are empty}
\usage{
is_missing_value(x)
}
\arguments{
\item{x}{An atomic vector}
}
\value{
A logical vector of the same length
}
\description{
NA, blank strings and the ICASA missing value (stored as "-99.0" or -99)
count as empty.
}
