% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{lang_to_iso}
\alias{lang_to_iso}
\title{Convert language column name to ISO code}
\usage{
lang_to_iso(language)
}
\arguments{
\item{language}{Either an ISO code or a display_names.csv column name}
}
\value{
ISO language code
}
\description{
Convert language column name to ISO code
}
