% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{lookup_property}
\alias{lookup_property}
\title{Look up a property descriptor from the registry}
\usage{
lookup_property(registry, prop_name, event_type = NULL, subtype = NULL)
}
\arguments{
\item{registry}{The property registry}

\item{prop_name}{Property name}

\item{event_type}{Event type const (or NULL for common)}

\item{subtype}{Subtype const (or NULL)}
}
\value{
The property descriptor, or NULL
}
\description{
Look up a property descriptor from the registry
}
