% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{make_required_label}
\alias{make_required_label}
\title{Create a label with a required asterisk indicator}
\usage{
make_required_label(label, required)
}
\arguments{
\item{label}{The label text}

\item{required}{Whether the field is required}
}
\value{
The label, optionally with a red asterisk appended
}
\description{
Create a label with a required asterisk indicator
}
