% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{render_array_table}
\alias{render_array_table}
\title{Render a schema array property as a table module placeholder}
\usage{
render_array_table(desc, ns, iso)
}
\arguments{
\item{desc}{The property descriptor of the array (e.g. planting_list)}

\item{ns}{Namespace function}

\item{iso}{ISO language code}
}
\value{
A tagList with the table module UI
}
\description{
Render a schema array property as a table module placeholder
}
