% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{render_event_panel}
\alias{render_event_panel}
\title{Render the panel for a single event type}
\usage{
render_event_panel(event_entry, pr, ns, iso, event_const)
}
\arguments{
\item{event_entry}{An event registry entry}

\item{pr}{Property registry}

\item{ns}{Namespace function}

\item{iso}{ISO language code}

\item{event_const}{The event type const value}
}
\value{
A tagList
}
\description{
Render the panel for a single event type
}
