% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{render_field}
\alias{render_field}
\title{Render the widget or table of a field, shown only when its x-ui condition
(if any) is met}
\usage{
render_field(desc, ns, iso)
}
\arguments{
\item{desc}{Property descriptor}

\item{ns}{Namespace function}

\item{iso}{ISO language code}
}
\value{
A Shiny tag
}
\description{
Render the widget or table of a field, shown only when its x-ui condition
(if any) is met
}
