% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_language.R
\name{replace_with_display_names}
\alias{replace_with_display_names}
\title{Replace code names with display names in an event data frame}
\usage{
replace_with_display_names(events_with_code_names, language)
}
\arguments{
\item{events_with_code_names}{The data frame with code names that should be
turned to display names}

\item{language}{The language of the display names}
}
\value{
A data frame of the same size but with entries with code names
  replaced with display names
}
\description{
Also replaces missingvals with "".
}
