% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{reset_input_fields}
\alias{reset_input_fields}
\title{Reset the value of input fields}
\usage{
reset_input_fields(session, fields_to_clear, exceptions = c(""))
}
\arguments{
\item{session}{The current Shiny session}

\item{fields_to_clear}{The names of the variables whose corresponding fields
should be cleared}

\item{exceptions}{Optional vector of variable names which should not be
cleared.}
}
\value{
None, used for side effects.
}
\description{
Reset the value of input fields
}
