% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{resolve_ref}
\alias{resolve_ref}
\title{Resolve a $ref pointer within the schema $defs}
\usage{
resolve_ref(obj, defs)
}
\arguments{
\item{obj}{A list that may contain a $ref key}

\item{defs}{The $defs section from the schema}
}
\value{
The resolved object
}
\description{
Resolve a $ref pointer within the schema $defs
}
