% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{rlapply}
\alias{rlapply}
\title{Recursively apply function to lists in a list}
\usage{
rlapply(x, fun, name_fun = NULL, ...)
}
\arguments{
\item{x}{The list of lists to apply the function to}

\item{fun}{The function to apply to lists}

\item{name_fun}{Function used to name the elements of the returned list.
Should take a list as argument and return the name}

\item{...}{arguments to pass to fun}
}
\value{
A one-level list where each element is the value fun returns for a
given list in x
}
\description{
Recursively apply a function to the elements of a list that
are themselves lists.
}
