% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema.R
\name{schema_get_choices}
\alias{schema_get_choices}
\title{Build a named choice vector for selectInput from schema choices}
\usage{
schema_get_choices(choices, language)
}
\arguments{
\item{choices}{A list of choice descriptors from extract_oneof_choices}

\item{language}{Language code or column name}
}
\value{
A named character vector suitable for selectInput choices
}
\description{
Build a named choice vector for selectInput from schema choices
}
