% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{schema_table_id}
\alias{schema_table_id}
\title{Module ID of the table for an array property}
\usage{
schema_table_id(desc)
}
\arguments{
\item{desc}{The property descriptor of the array}
}
\value{
The table module ID
}
\description{
Module ID of the table for an array property
}
