% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_language.R
\name{set_login_language}
\alias{set_login_language}
\title{This function sets the labels on the shinymanager login UI}
\usage{
set_login_language(language)
}
\arguments{
\item{language}{The language which should be displayed (either
"disp_name_fin" or disp_name_eng)}
}
\description{
This function sets the labels on the shinymanager login UI
}
