% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{update_ui_element}
\alias{update_ui_element}
\title{Update value, label etc. of a UI element.}
\usage{
update_ui_element(session, code_name, value = NULL, clear_value = FALSE, ...)
}
\arguments{
\item{session}{Current shiny session}

\item{code_name}{The code name of the UI element to update}

\item{value}{An atomic vector holding the desired value of the UI element.}

\item{clear_value}{If set to TRUE, the value of the element is cleared.}

\item{...}{Additional arguments to pass to Shiny's update functions.}
}
\description{
Determines the type of the element and updates its value using shiny's update
  functions.
}
