% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{valid_dateRangeInput}
\alias{valid_dateRangeInput}
\title{Check whether the value of a dateRangeInput is valid}
\usage{
valid_dateRangeInput(value)
}
\value{
TRUE if value is valid, FALSE if not
}
\description{
Both dates need to be supplied for the value to be considered
  valid
}
