% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{write_json_file}
\alias{write_json_file}
\title{Write a given event list to a json file}
\usage{
write_json_file(
  site,
  block,
  event_list,
  rotation_list,
  base_folder = json_file_base_folder()
)
}
\arguments{
\item{site}{The site of the events}

\item{block}{The block of the events}

\item{event_list}{The list of events to write to the events.json file}

\item{base_folder}{Included for testing reasons, the default value should
otherwise be used}
}
\description{
The function will overwrite the current events.json file and replace it with
one generated from the supplied list of events
}
