% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{build_structure_lookup_list}
\alias{build_structure_lookup_list}
\title{Build lookup list for UI elements}
\usage{
build_structure_lookup_list()
}
\value{
The lookup list.
}
\description{
Build a list where the names are the code names of UI elements
and the values are the corresponding element structures (lists) found in 
ui_structure.json. Now only contains app chrome elements.
}
