% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{copy_file}
\alias{copy_file}
\title{Copy a file related to an event and name it appropriately}
\usage{
copy_file(
  orig_filepath,
  variable_name,
  site,
  block,
  date,
  filepath_is_relative = FALSE,
  delete_original = FALSE,
  base_folder = json_file_base_folder()
)
}
\arguments{
\item{orig_filepath}{The path of the file to copy}

\item{variable_name}{Which variable is this file for? E.g. canopeo_image}

\item{site}{The site where the event took place}

\item{block}{The block where the event took place}

\item{date}{The day of the event as a character string, format yyyy-mm-dd}

\item{filepath_is_relative}{If TRUE, json_file_base_folder will be added}

\item{delete_original}{Should the original file be deleted after copying?}

\item{base_folder}{Included for testing reasons}
}
\value{
A path to the new location of the file relative to events.json.
}
\description{
When a file (image) is uploaded through a fileInput widget, it is saved to a
temporary folder. This function copies that file to an appropriate directory
and name. The file does not have to be originally in a temporary folder —
any file path is valid. This allows the function to also be used when cloning
an event and its associated images need to be duplicated.
}
\details{
The new file name has the format
  `yyyy-mm-dd_site_block_variable_name_#.ext` where `#` is an incrementing
  number (0, 1, 2, ...) to ensure uniqueness within the target folder.
}
