% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{copy_file}
\alias{copy_file}
\title{Copy a file related to an event and name it appropriately}
\usage{
copy_file(
  orig_filepath,
  variable_name,
  site,
  block,
  date,
  filepath_is_relative = FALSE,
  delete_original = FALSE,
  base_folder = json_file_base_folder()
)
}
\arguments{
\item{orig_filepath}{The path of the file to copy}

\item{variable_name}{Which variable is this file for? E.g. canopeo_image}

\item{site}{The site where the event took place}

\item{block}{The block where the event took place}

\item{date}{The day of the event as a character string, the format must be
yyyy-mm-dd}

\item{filepath_is_relative}{If TRUE, json_file_base_folder will be added to
the beginning of filepath}

\item{delete_original}{Should the original file be deleted after copying?}

\item{base_folder}{Included for testing reasons, the default value should
otherwise be used}
}
\value{
A path to the new location of the file relative to the events.json
  file.
}
\description{
When a file (image) is uploaded through a fileInput widget, it is saved to a
temporary folder. This function copies that file to an appropriate directory
and name. The file does not have to be originally in a temporary
folder, any file path is ok. Therefore this function can also be used e.g.
when cloning and event and the images associated with it need to be
duplicated.
}
\details{
The name will be of the format 
yyyy-mm-dd_site_block_variable_name_# where # is a number (0, 1, 2, ...) to
ensure that files have unique names.
}
