% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_json.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file}
\usage{
delete_file(filepath, site = NULL, block = NULL, filepath_relative = FALSE)
}
\arguments{
\item{filepath}{The path to the file which should be deleted.}

\item{site}{The site where the event took place}

\item{block}{The block where the event took place}

\item{filepath_relative}{Set to TRUE and supply site and block if filepath is
relative to the events.json file. This allows the function to figure out
the correct path to the file.}
}
\description{
Delete the file with the path filepath. Used to delete files (images)
associated with events, e.g. canopeo_image
}
