% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{get_selectInput_choices}
\alias{get_selectInput_choices}
\title{Find the choices for a selectInput given its code name}
\usage{
get_selectInput_choices(selectInput_code_name, language)
}
\arguments{
\item{selectInput_code_name}{The code name of the selectInput}

\item{language}{The language to show the options in. This will be passed to
get_disp_name}
}
\value{
A vector of choices (code names). If language was supplied, the names
  will be the names of the vector.
}
\description{
Find the choices for a selectInput given its code name
}
