% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_event_list.R
\name{mod_event_list_server}
\alias{mod_event_list_server}
\title{event_list Server Functions}
\usage{
mod_event_list_server(id, events, language, site)
}
\arguments{
\item{id}{The id of the corresponding UI element}

\item{events}{A reactive expression holding a list of events to display in the event list}

\item{language}{A reactive expression holding the current UI language}

\item{site}{A reactive expression holding the current site name

Update year choices in event list filter

Adds as choices all the years for which events have been recorded
Update block choices in event list filter

Add all blocks of the current site as choices
Update activity choices in event list filter

Only used when the language is changed.}
}
\description{
event_list Server Functions
}
