% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{normalize_legacy_event}
\alias{normalize_legacy_event}
\title{Normalize a legacy event to use schema property names}
\usage{
normalize_legacy_event(event)
}
\arguments{
\item{event}{An event list}
}
\value{
The event with legacy names mapped to schema names
}
\description{
Normalize a legacy event to use schema property names
}
