% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{read_json_file}
\alias{read_json_file}
\title{Read the events from the events.json file}
\usage{
read_json_file(site, block, base_folder = json_file_base_folder())
}
\arguments{
\item{site}{The site to read from}

\item{block}{The block to read from}

\item{base_folder}{Included for testing reasons, the default value should
otherwise be used}
}
\value{
A list of events, which are themselves lists. If the corresponding
  file does not exist or there are no events, returns an empty list.
}
\description{
Reads the events from the events.json file specific to this site and block
combination and returns as a list of events.
}
