% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{reset_input_fields}
\alias{reset_input_fields}
\title{Reset the value of input fields}
\usage{
reset_input_fields(session, fields_to_clear, exceptions = c(""))
}
\arguments{
\item{session}{The current Shiny session}

\item{fields_to_clear}{The names of the variables whose corresponding fields
should be cleared}

\item{exceptions}{Optional vector of variable names which should not be
cleared. This is useful if fields_to_clear is supplied with all variable
names but there are a few that should not be cleared.}
}
\value{
None, used for side effects.
}
\description{
Set the specified input fields to their default empty values.
}
\note{
This doesn't reset the tables (e.g. harvest_crop_table) -- they reset 
  themselves every time they become hidden.
}
