% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{update_schema_value}
\alias{update_schema_value}
\title{Update a schema-driven widget value (for populating forms)}
\usage{
update_schema_value(session, prop_name, desc, value)
}
\arguments{
\item{session}{Shiny session}

\item{prop_name}{Property name}

\item{desc}{Property descriptor}

\item{value}{The value to set}
}
\description{
Update a schema-driven widget value (for populating forms)
}
