% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_schema_ui.R
\name{update_schema_widget}
\alias{update_schema_widget}
\title{Update a single schema widget's label and choices}
\usage{
update_schema_widget(session, prop_name, desc, iso, input)
}
\description{
Update a single schema widget's label and choices
}
